LayoutTests/fast/filesystem/filesystem-missing-arguments.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test how file system methods react to too few arguments.');
shouldThrow("webkitRequestFileSystem()");
shouldThrow("webkitRequestFileSystem(TEMPORARY)");
shouldThrow("webkitResolveLocalFileSystemURL()");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
